Large refactor of http/http2 request cycles to accomodate http2, httpfactory for protocol negotiation.
